INtroduce middleware to capture exceptions
