+++ lib/redux/middleware/sentry_api_middleware.dart
@@ -1,7 +1,4 @@
-import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:package_info/package_info.dart';
 import 'package:redux/redux.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../api/sentry_api.dart';
 import '../../types/cursor.dart';
@@ -36,8 +33,8 @@ class SentryApiMiddleware extends MiddlewareClass<AppState> {
 
             final nextCursor = action.pagination
                 ? currentCursor == null
-                  ? Cursor(10, 0, 0)
-                  : Cursor(10, currentCursor.offset + 1, 0)
+                ? Cursor(10, 0, 0)
+                : Cursor(10, currentCursor.offset + 1, 0)
                 : null;
 
             final projects = await api.projects(organization.slug, nextCursor);
@@ -73,7 +70,7 @@ class SentryApiMiddleware extends MiddlewareClass<AppState> {
 
           for (final organizationSlug in action.projectsByOrganizationSlug.keys) {
             final projectsToFetch = (action.projectsByOrganizationSlug[organizationSlug] ?? [])
-              .where((element) => element.latestRelease != null); // Only fetch when there is a release
+                .where((element) => element.latestRelease != null); // Only fetch when there is a release
 
             for (final projectToFetch in projectsToFetch) {
               final project = await api.project(
@@ -117,8 +114,8 @@ class SentryApiMiddleware extends MiddlewareClass<AppState> {
         final api = SentryApi(store.state.globalState.authToken);
         try {
           final List<Group> issues = await api.issues(
-              organizationSlug: action.organizationSlug,
-              projectSlug: action.projectSlug,
+            organizationSlug: action.organizationSlug,
+            projectSlug: action.projectSlug,
           );
           store.dispatch(
               FetchIssuesSuccessAction(action.projectSlug, issues)
@@ -150,19 +147,19 @@ class SentryApiMiddleware extends MiddlewareClass<AppState> {
         final api = SentryApi(store.state.globalState.authToken);
         try {
           final sessions = await api.sessions(
-            organizationSlug: action.organizationSlug,
-            projectId: action.projectId,
-            fields: [SessionGroup.sumSessionKey, SessionGroup.countUniqueUsersKey],
-            groupBy: SessionGroupBy.sessionStatusKey
+              organizationSlug: action.organizationSlug,
+              projectId: action.projectId,
+              fields: [SessionGroup.sumSessionKey, SessionGroup.countUniqueUsersKey],
+              groupBy: SessionGroupBy.sessionStatusKey
           );
 
           final sessionsBefore = await api.sessions(
-            organizationSlug: action.organizationSlug,
-            projectId: action.projectId,
-            fields: [SessionGroup.sumSessionKey, SessionGroup.countUniqueUsersKey],
-            groupBy: SessionGroupBy.sessionStatusKey,
-            statsPeriodStart: '48h',
-            statsPeriodEnd: '24h'
+              organizationSlug: action.organizationSlug,
+              projectId: action.projectId,
+              fields: [SessionGroup.sumSessionKey, SessionGroup.countUniqueUsersKey],
+              groupBy: SessionGroupBy.sessionStatusKey,
+              statsPeriodStart: '48h',
+              statsPeriodEnd: '24h'
           );
 
           store.dispatch(
@@ -200,11 +197,11 @@ class SentryApiMiddleware extends MiddlewareClass<AppState> {
           final fortyEightHoursAgo = now.add(Duration(hours: -48));
 
           final apdex = await api.apdex(
-            apdexThreshold: action.apdexThreshold,
-            organizationSlug: action.organizationSlug,
-            projectId: action.projectId,
-            start: twentyFourHoursAgo,
-            end: now
+              apdexThreshold: action.apdexThreshold,
+              organizationSlug: action.organizationSlug,
+              projectId: action.projectId,
+              start: twentyFourHoursAgo,
+              end: now
           );
 
           final apdexBefore = await api.apdex(
@@ -228,30 +225,5 @@ class SentryApiMiddleware extends MiddlewareClass<AppState> {
     } else {
       next(action);
     }
-  }
-}
-
-class LocalStorageMiddleware extends MiddlewareClass<AppState> {
-  LocalStorageMiddleware(this.preferences, this.secureStorage);
-
-  final SharedPreferences preferences;
-  final FlutterSecureStorage secureStorage;
-
-  @override
-  dynamic call(Store<AppState> store, dynamic action, NextDispatcher next) async {
-    if (action is RehydrateAction) {
-      final String authToken = await secureStorage.read(key: 'authToken');
-      final packageInfo = await PackageInfo.fromPlatform();
-      final version = 'Version ${packageInfo.version} (${packageInfo.buildNumber})';
-      store.dispatch(RehydrateSuccessAction(authToken, version));
-    }
-    if (action is LoginAction) {
-      await secureStorage.write(key: 'authToken', value: action.authToken);
-    }
-    if (action is LogoutAction) {
-      await secureStorage.delete(key: 'authToken');
-      await preferences.clear();
-    }
-    next(action);
   }
 }
